Move limb methods to `FqVectorBase`
