+++ ext/crates/fp/src/vector/impl_fqslice.rs
@@ -4,7 +4,7 @@ use super::{
     inner::{FqSlice, FqVector},
     iter::{FqVectorIterator, FqVectorNonZeroIterator},
 };
-use crate::{constants, field::Field, limb::Limb};
+use crate::field::Field;
 
 // Public methods
 
@@ -69,56 +69,6 @@ impl<'a, F: Field> FqSlice<'a, F> {
             new.as_slice_mut().assign(self);
         }
         new
-    }
-}
-
-// Limb methods
-impl<F: Field> FqSlice<'_, F> {
-    #[inline]
-    pub(super) fn offset(&self) -> usize {
-        let bit_length = self.fq().bit_length();
-        let entries_per_limb = self.fq().entries_per_limb();
-        (self.start() % entries_per_limb) * bit_length
-    }
-
-    #[inline]
-    pub(super) fn limb_range(&self) -> std::ops::Range<usize> {
-        self.fq().range(self.start(), self.end())
-    }
-
-    /// This function underflows if `self.end() == 0`, which happens if and only if we are taking a
-    /// slice of width 0 at the start of an `FpVector`. This should be a very rare edge case.
-    /// Dealing with the underflow properly would probably require using `saturating_sub` or
-    /// something of that nature, and that has a nontrivial (10%) performance hit.
-    #[inline]
-    pub(super) fn limb_range_inner(&self) -> std::ops::Range<usize> {
-        let range = self.limb_range();
-        (range.start + 1)..(usize::max(range.start + 1, range.end - 1))
-    }
-
-    #[inline(always)]
-    pub(super) fn min_limb_mask(&self) -> Limb {
-        !0 << self.offset()
-    }
-
-    #[inline(always)]
-    pub(super) fn max_limb_mask(&self) -> Limb {
-        let num_entries = 1 + (self.end() - 1) % self.fq().entries_per_limb();
-        let bit_max = num_entries * self.fq().bit_length();
-
-        (!0) >> (constants::BITS_PER_LIMB - bit_max)
-    }
-
-    #[inline(always)]
-    pub(super) fn limb_masks(&self) -> (Limb, Limb) {
-        if self.limb_range().len() == 1 {
-            (
-                self.min_limb_mask() & self.max_limb_mask(),
-                self.min_limb_mask() & self.max_limb_mask(),
-            )
-        } else {
-            (self.min_limb_mask(), self.max_limb_mask())
-        }
     }
 }
